refactor: clean up http module
