+++ src/http/mod.rs
@@ -2,10 +2,14 @@ mod client;
 mod data_loader;
 
 mod data_loader_request;
+mod handler;
+mod http_1;
+mod http_2;
 mod method;
 mod request_context;
 mod response;
 mod server;
+mod server_config;
 mod server_context;
 
 use std::time::Duration;
@@ -14,12 +18,15 @@ use cache_control::CacheControl;
 pub use client::*;
 pub use data_loader::*;
 pub use data_loader_request::*;
+pub use handler::{handle_batch_request, handle_single_request};
 use hyper::header::CACHE_CONTROL;
 pub use method::Method;
 pub use request_context::RequestContext;
 pub use response::*;
-pub use server::{handle_batch_request, handle_single_request, start_server};
+pub use server::start_server;
 pub use server_context::ServerContext;
+
+use self::server_config::ServerConfig;
 
 pub fn max_age(res: &Response) -> Option<Duration> {
   let header = res.headers.get(CACHE_CONTROL)?;
@@ -41,6 +48,14 @@ pub fn min_ttl<'a>(res_vec: impl Iterator<Item = &'a Response>) -> i32 {
     }
   }
   min
+}
+
+fn log_launch(sc: &ServerConfig) {
+  let addr = sc.addr().to_string();
+  log::info!("🚀 Tailcall launched at [{}] over {}", addr, sc.http_version());
+  if sc.graphiql() {
+    log::info!("🌍 Playground: https://{}", addr);
+  }
 }
 
 #[cfg(test)]
